🐛  move ghost CMS to api folder
