--- apps/website/pages/blog/ghostCMS/tags.tsx
@@ -1,5 +1,0 @@
-import { ghostApi } from '.'
-
-export const getTags = async () => {
-  return await ghostApi.tags.browse()
-}
