+++ apps/website/pages/blog/index.tsx
@@ -2,15 +2,15 @@ import { NextPage } from 'next'
 import React from 'react'
 import { typography } from '@blockfint/website/styles/typography'
 import { createGlobalStyle } from 'styled-components'
-import { getAllPosts } from './ghostCMS/index'
+import { getAllPosts } from '../api/ghostCMS/index'
 import { TestContents } from '../../containers/blog'
+import { typeOf } from 'react-is'
 
 const Global = createGlobalStyle`
 body{
   ${typography}
 }
 `
-type Post = {}
 
 export const getStaticProps = async ({ params }) => {
   const AllBlogs = await getAllPosts()
@@ -20,7 +20,8 @@ export const getStaticProps = async ({ params }) => {
   }
 }
 
-const BlogPage: React.FC<{ blogs: Post[] }> = (props) => {
+const BlogPage: React.FC = (props) => {
+  console.log(props)
   return (
     <>
       <Global />
